+++ src/main/resources/benchmarks/twitter/ddl-nuodb.sql
@@ -2,7 +2,7 @@
 
 DROP TABLE IF EXISTS user_profiles CASCADE;
 CREATE TABLE user_profiles (
-  uid int NOT NULL,
+  uid bigint NOT NULL,
   name string DEFAULT NULL,
   email string DEFAULT NULL,
   partitionid int DEFAULT NULL,
@@ -15,15 +15,15 @@ CREATE INDEX IDX_USER_PARTITION ON user_profiles (partitionid);
 
 DROP TABLE IF EXISTS followers CASCADE;
 CREATE TABLE followers (
-  f1 int NOT NULL REFERENCES user_profiles (uid),
-  f2 int NOT NULL REFERENCES user_profiles (uid),
+  f1 bigint NOT NULL REFERENCES user_profiles (uid),
+  f2 bigint NOT NULL REFERENCES user_profiles (uid),
   PRIMARY KEY (f1,f2)
 );
 
 DROP TABLE IF EXISTS follows CASCADE;
 CREATE TABLE follows (
-  f1 int NOT NULL REFERENCES user_profiles (uid),
-  f2 int NOT NULL REFERENCES user_profiles (uid),
+  f1 bigint NOT NULL REFERENCES user_profiles (uid),
+  f2 bigint NOT NULL REFERENCES user_profiles (uid),
   PRIMARY KEY (f1,f2)
 );
 
@@ -31,7 +31,7 @@ CREATE TABLE follows (
 DROP TABLE IF EXISTS tweets CASCADE;
 CREATE TABLE tweets (
   id bigint NOT NULL,
-  uid int NOT NULL REFERENCES user_profiles (uid),
+  uid bigint NOT NULL REFERENCES user_profiles (uid),
   text string NOT NULL,
   createdate datetime DEFAULT NULL,
   PRIMARY KEY (id)
@@ -42,7 +42,7 @@ CREATE INDEX IDX_TWEETS_UID ON tweets (uid);
 DROP TABLE IF EXISTS added_tweets CASCADE;
 CREATE TABLE added_tweets (
   id bigint generated always as identity,
-  uid int NOT NULL REFERENCES user_profiles (uid),
+  uid bigint NOT NULL REFERENCES user_profiles (uid),
   text string NOT NULL,
   createdate datetime DEFAULT NULL
 --  PRIMARY KEY (id)
